Add useIsLoggedIn hook to have a login status for global use and refactor Routes component
